requirements.txt
# django
django == 4.1.4
psycopg2 == 2.9.5
pillow == 9.3.0
django-split-settings==1.2.0
toml==0.10.2